addition of few modifiers

+++ Ball.cpp
@@ -4,7 +4,7 @@ Ball::Ball(const float radius, const float delta_value, const sf::Vector2u&& win
     : sf::CircleShape(radius), x_delta(delta_value), y_delta(delta_value) {
     this->setFillColor(sf::Color::White);
     this->setOrigin(sf::Vector2f(radius, radius));
-    this->setPosition(sf::Vector2f(window_size.x / 2.f, window_size.y / 2.f));
+    this->setPosition(sf::Vector2f(static_cast<float>(window_size.x) / 2.f, static_cast<float>(window_size.y) / 2.f));
 }
 
 void Ball::paddleCollision(const Paddle& player_one, const Paddle& player_two) {
@@ -20,13 +20,13 @@ void Ball::paddleCollision(const Paddle& player_one, const Paddle& player_two) {
 
 bool Ball::wallCollision(const sf::Vector2u& window_size) {
     if (this->getPosition().x - this->getRadius() < 0 ||
-        this->getPosition().x + this->getRadius() > window_size.x) {
-        this->setPosition(sf::Vector2f(window_size.x / 2.f, window_size.y / 2.f));
+        this->getPosition().x + this->getRadius() > static_cast<float>(window_size.x)) {
+        this->setPosition(sf::Vector2f(static_cast<float>(window_size.x) / 2.f, static_cast<float>(window_size.y) / 2.f));
         x_delta = -x_delta;
         }
 
     return (this->getPosition().y - this->getRadius() < 0 ||
-        this->getPosition().y + this->getRadius() > window_size.y);
+        this->getPosition().y + this->getRadius() > static_cast<float>(window_size.y));
 }
 
 void Ball::move() {
